Limpieza de codigo realizada
